* 'develop' of https://github.com/project-pet-meeting/repo-be: (29 commits)
  Entity 연관관계 재설정
  불필요한 import 문 제거
  회원가입 시 이미지 업로드 기능 추가 #2
  카카오 소셜 로그인 일부 코드 수정 #2
  네이버 소셜 로그인 구현 #2
  게시글 내 카테고리 필드 Enum형으로 변경 #3
  전역 에러, 예외 처리 수정 및 코드 추가 / post 예외 처리 추가 #3
  이미지 업로드 유효성 검사 추가(파일 존재 유무, 파일 확장자) #3
  1. 게시글 작성 회원 정보 추가 2. 게시글 수정, 삭제 권한 확인 추가 #3
  카카오 소셜 로그인 구현 #2
  Email 인증 추가 #2
  전역 예외처리 및 일부 코드 수정
  ASCII 도입
  Post CURD 기능 구현 코드 수정 #3
  게시글 관련 예외 처리 추가 #3
  게시글 CRUD + 이미지 업로드 #3
  Post CRUD 기능 구현 수정  #3
  link 추가 및 Test 코드 수정 #2
  AWS RDS 연결 설정
  로그아웃 기능 구현 완료 #2
  ...

+++ pet-meeting/src/main/java/sideproject/petmeeting/post/service/PostService.java
@@ -10,14 +10,13 @@ import org.springframework.web.multipart.MultipartFile;
 import sideproject.petmeeting.common.exception.BusinessException;
 import sideproject.petmeeting.common.exception.ErrorCode;
 import sideproject.petmeeting.common.S3Uploader;
+import sideproject.petmeeting.member.domain.Member;
 import sideproject.petmeeting.post.Repository.PostRepository;
 import sideproject.petmeeting.post.domain.Post;
 import sideproject.petmeeting.post.dto.PostRequestDto;
 import sideproject.petmeeting.post.dto.PostResponseDto;
 
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
 
 
 @RequiredArgsConstructor
@@ -31,12 +30,19 @@ public class PostService {
      * @param postRequestDto : 게시글 작성에 필요한 값
      */
     @Transactional
-    public Post createPost(PostRequestDto postRequestDto, MultipartFile image) throws IOException {
-        String imageUrl = s3Uploader.upload(image, "/post/imgae");
-        Post post = new Post(postRequestDto, imageUrl);
-        postRepository.save(post);
+    public Post createPost(PostRequestDto postRequestDto, MultipartFile image, Member member) throws IOException {
+        String imageUrl = s3Uploader.upload(image, "/post/image");
 
-        return post;
+        Post post = Post.builder()
+                .category(postRequestDto.getCategory())
+                .title(postRequestDto.getTitle())
+                .content(postRequestDto.getContent())
+                .imageUrl(imageUrl)
+                .member(member)
+                .build();
+
+        return postRepository.save(post);
+
     }
 
 
@@ -68,24 +74,25 @@ public class PostService {
      * 게시글 수정
      * @param postId : 수정할 게시글 id
      * @param postRequestDto : 수정할 게시글
-     * @param image
+     * @param image : 수정할 이미지 파일
      * @return
      * @throws IOException
      */
     @Transactional
-    public Post updatePost(Long postId, PostRequestDto postRequestDto, MultipartFile image) throws IOException {
+    public Post updatePost(Long postId, PostRequestDto postRequestDto, MultipartFile image, Member member) throws IOException {
         Post post = postRepository.findById(postId).orElseThrow(
                 () -> new BusinessException("존재하지 않는 게시글 id 입니다.", ErrorCode.POST_NOT_EXIST)
         );
 
+        if (!post.getMember().getId().equals(member.getId())) {
+            throw new BusinessException("수정 권한이 없습니다.", ErrorCode.HANDLE_ACCESS_DENIED);
+        }
+
         String imageUrl = post.getImageUrl();
-        String postImage = "";
-        List<String> imgPaths = null;
 
         // 이미지 존재 시 삭제 후 업로드
         if (imageUrl != null) {
-            String deleteUrl = imageUrl.substring(imageUrl.indexOf("post/image"));
-            s3Uploader.deleteImage(deleteUrl);
+            s3Uploader.deleteImage(imageUrl);
         }
 
         imageUrl = s3Uploader.upload(image, "/post/image");
@@ -102,29 +109,23 @@ public class PostService {
      * @throws IOException : 삭제할 게시글의 image 파일명 인코딩 예외 처리, UnsupportedEncodingException
      */
     @Transactional
-    public void postDelete(Long postId) throws IOException{
+    public void postDelete(Long postId, Member member) throws IOException{
         Post post = postRepository.findById(postId).orElseThrow(
                 () -> new BusinessException("존재하지 않는 게시글 id 입니다.", ErrorCode.POST_NOT_EXIST)
         );
 
+        if (!post.getMember().getId().equals(member.getId())) {
+            throw new BusinessException("삭제 권한이 없습니다.", ErrorCode.HANDLE_ACCESS_DENIED);
+        }
+
         String imageUrl = post.getImageUrl();
+
         if (imageUrl != null) {
-            String deleteUrl = imageUrl.substring(imageUrl.indexOf("post/image"));
-            s3Uploader.deleteImage(deleteUrl);
+            s3Uploader.deleteImage(imageUrl);
         }
 
         postRepository.deleteById(postId);
 
-
-   }
-
-
-
-
-
-
-
-
-
+    }
 
 }
